Move extra methods to a utility file. And styles to css no point in having big strings in memory the whole time. plus its a lot easier to edit a css file.
