+++ src/constants.ts
@@ -1,6 +1,7 @@
 export const STORY_BOOK_BLOCK = `import { Meta } from "@storybook/blocks";
-
+import './style.css';
 
 `;
 
 export const typelit = '&lcub;...&rcub;';
+
